Add support for escaped KVC characters.
This implementation is really rough. There is likely room to iterate.

What this solution does is provides support for escaping the following
characters in patterns:

- .
- @
- :
- \

These characters may be escaped using a double backslash. We use a double
backslash rather than a single backslash because a single backslash acts
as an Objective-C string literal character escape which is not what we
want here (and will raise compiler warnings due to unrecognized escape
characters). The double backslash allows us to programmatically replace
the above characters while matching the pattern.

An example pattern that requires escaping:

    /:userid.json

Without character escaping, this pattern would be interpreted as fetching
the userid property and then fetching the json property of that userid. This would
throw a KVC exception. When going outbound, a string of the form "/3.json" would end
up passing "3.json" as the parameter, rather than the desired "3".

The escaped version of this string would look like this:

    /:userid\\.json

This allows us to create a string from an object's userid property of the form
"/3.json". It also allows us to extract the correct value from a string of the
form "/3.json", providing the value "3" to the first parameter of the selector.